Add hangup function to Hikconnect API
A new asynchronous function 'hangup' was added to the hikconnect/api.py. The purpose of this function is to send a hangup request to a specific device, using the device's unique serial number. This serial number can be obtained through either the 'get_devices()' or 'get_cameras()' functions. Changes have also been made in tests/test_manual.py to accommodate for this new function.